Updated to PHP 7

+++ scripts/homestead.rb
@@ -89,9 +89,11 @@ class Homestead
 
     # Configure The Public Key For SSH Access
     if settings.include? 'authorize'
-      config.vm.provision "shell" do |s|
-        s.inline = "echo $1 | grep -xq \"$1\" /home/vagrant/.ssh/authorized_keys || echo $1 | tee -a /home/vagrant/.ssh/authorized_keys"
-        s.args = [File.read(File.expand_path(settings["authorize"]))]
+      if File.exists? File.expand_path(settings["authorize"])
+        config.vm.provision "shell" do |s|
+          s.inline = "echo $1 | grep -xq \"$1\" /home/vagrant/.ssh/authorized_keys || echo $1 | tee -a /home/vagrant/.ssh/authorized_keys"
+          s.args = [File.read(File.expand_path(settings["authorize"]))]
+        end
       end
     end
 
@@ -172,7 +174,7 @@ class Homestead
     if settings.has_key?("variables")
       settings["variables"].each do |var|
         config.vm.provision "shell" do |s|
-          s.inline = "echo \"\nenv[$1] = '$2'\" >> /etc/php5/fpm/php-fpm.conf"
+          s.inline = "echo \"\nenv[$1] = '$2'\" >> /etc/php/7.0/fpm/php-fpm.conf"
           s.args = [var["key"], var["value"]]
         end
 
@@ -183,7 +185,7 @@ class Homestead
       end
 
       config.vm.provision "shell" do |s|
-        s.inline = "service php5-fpm restart"
+        s.inline = "service php7.0-fpm restart"
       end
     end
